requirements.txt
dotenv
google-generativeai
streamlit